Move ScreenStack to render module
