+++ tests/screen_stack_test.py
@@ -21,8 +21,9 @@
 
 
 import unittest
+
+from simpleline.render.screen_stack import ScreenStack, ScreenData, ScreenStackEmptyException
 from simpleline.render.ui_screen import UIScreen
-from simpleline.screen_stack import ScreenStack, ScreenData, ScreenStackEmptyException
 
 
 class ScreenStack_TestCase(unittest.TestCase):
